chore(ET-5848): Use HashMap for measurement API

+++ core/core-api/src/services/proton/measurements/responses.rs
@@ -4,7 +4,7 @@ use serde::Deserialize;
 #[allow(dead_code)]
 pub struct PostMeasurementEventResponse {
     #[serde(default)]
-    pub session_start_ms: Option<i64>,
+    pub session_start_ms: Option<u128>,
     // Note: This field exists in response specs but we decided not to use it.
     // To prevent accidental use I commented it but I'm leaving this comment
     // so its clear this is intentional and not a mistake.
